chore: refactor in-memory key store implementation
